feat(ui): move NUMA resources view from Resources to VMs tab
