+++ ui/src/app/kvm/components/LXDHostVMs/NumaResources/NumaResources.tsx
