+++ ui/src/app/kvm/components/LXDHostVMs/NumaResources/NumaResourcesCard/_index.scss
