+++ ui/src/app/kvm/components/LXDHostVMs/NumaResources/NumaResourcesCard/index.ts
