+++ ui/src/app/kvm/components/LXDHostVMs/NumaResources/_index.scss
