+++ ui/src/app/kvm/components/LXDHostVMs/NumaResources/index.ts
